+++ es5/videojs-dash.js
@@ -1,27 +1,51 @@
-import window from 'global/window';
-import videojs from 'video.js';
-import dashjs from 'dashjs';
-import setupAudioTracks from './setup-audio-tracks';
-import setupTextTracks from './setup-text-tracks';
-import document from 'global/document';
+'use strict';
+
+exports.__esModule = true;
+
+var _window = require('global/window');
+
+var _window2 = _interopRequireDefault(_window);
+
+var _video = require('video.js');
+
+var _video2 = _interopRequireDefault(_video);
+
+var _dashjs = require('dashjs');
+
+var _dashjs2 = _interopRequireDefault(_dashjs);
+
+var _setupAudioTracks = require('./setup-audio-tracks');
+
+var _setupAudioTracks2 = _interopRequireDefault(_setupAudioTracks);
+
+var _setupTextTracks = require('./setup-text-tracks');
+
+var _setupTextTracks2 = _interopRequireDefault(_setupTextTracks);
+
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { 'default': obj }; }
+
+function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
 
 /**
  * videojs-contrib-dash
  *
  * Use Dash.js to playback DASH content inside of Video.js via a SourceHandler
  */
-class Html5DashJS {
-  constructor(source, tech, options) {
+var Html5DashJS = function () {
+  function Html5DashJS(source, tech, options) {
+    var _this = this;
+
+    _classCallCheck(this, Html5DashJS);
+
     // Get options from tech if not provided for backwards compatibility
     options = options || tech.options_;
 
-    this.player = videojs(options.playerId);
+    this.player = (0, _video2['default'])(options.playerId);
     this.player.dash = this.player.dash || {};
 
     this.tech_ = tech;
     this.el_ = tech.el();
     this.elParent_ = this.el_.parentNode;
-    this.hasFiniteDuration_ = false;
 
     // Do nothing if the src is falsey
     if (!source.src) {
@@ -34,54 +58,45 @@ class Html5DashJS {
     tech.isReady_ = false;
 
     if (Html5DashJS.updateSourceData) {
-      videojs.log.warn('updateSourceData has been deprecated.' +
-        ' Please switch to using hook("updatesource", callback).');
+      _video2['default'].log.warn('updateSourceData has been deprecated.' + ' Please switch to using hook("updatesource", callback).');
       source = Html5DashJS.updateSourceData(source);
     }
 
     // call updatesource hooks
-    Html5DashJS.hooks('updatesource').forEach((hook) => {
+    Html5DashJS.hooks('updatesource').forEach(function (hook) {
       source = hook(source);
     });
 
-    const manifestSource = source.src;
-
+    var manifestSource = source.src;
     this.keySystemOptions_ = Html5DashJS.buildDashJSProtData(source.keySystemOptions);
 
-    this.player.dash.mediaPlayer = dashjs.MediaPlayer().create();
+    this.player.dash.mediaPlayer = _dashjs2['default'].MediaPlayer().create();
 
     this.mediaPlayer_ = this.player.dash.mediaPlayer;
 
-    // For whatever reason, we need to call setTextDefaultEnabled(false) to get
-    // VTT captions to show, even though we're doing virtually the same thing
-    // in setup-text-tracks.js
-    this.mediaPlayer_.setTextDefaultEnabled(false);
-
     // Log MedaPlayer messages through video.js
     if (Html5DashJS.useVideoJSDebug) {
-      videojs.log.warn('useVideoJSDebug has been deprecated.' +
-        ' Please switch to using hook("beforeinitialize", callback).');
+      _video2['default'].log.warn('useVideoJSDebug has been deprecated.' + ' Please switch to using hook("beforeinitialize", callback).');
       Html5DashJS.useVideoJSDebug(this.mediaPlayer_);
     }
 
     if (Html5DashJS.beforeInitialize) {
-      videojs.log.warn('beforeInitialize has been deprecated.' +
-        ' Please switch to using hook("beforeinitialize", callback).');
+      _video2['default'].log.warn('beforeInitialize has been deprecated.' + ' Please switch to using hook("beforeinitialize", callback).');
       Html5DashJS.beforeInitialize(this.player, this.mediaPlayer_);
     }
 
-    Html5DashJS.hooks('beforeinitialize').forEach((hook) => {
-      hook(this.player, this.mediaPlayer_);
+    Html5DashJS.hooks('beforeinitialize').forEach(function (hook) {
+      hook(_this.player, _this.mediaPlayer_);
     });
 
     // Must run controller before these two lines or else there is no
     // element to bind to.
     this.mediaPlayer_.initialize();
 
-    this.timeUpdated = (event) => {
-        this.player.duration(this.duration());
-        this.tech_.trigger({ type: 'timeupdate', target: this.tech_, manuallyTriggered: true });
-    }
+    this.timeUpdated = function (event) {
+      _this.player.duration(_this.duration());
+      _this.tech_.trigger({ type: 'timeupdate', target: _this.tech_, manuallyTriggered: true });
+    };
     // Retrigger a dash.js-specific error event as a player error
     // See src/streaming/utils/ErrorHandler.js in dash.js code
     // Handled with error (playback is stopped):
@@ -93,32 +108,22 @@ class Html5DashJS {
     // Not handled:
     // - timedTextError (video can still play)
     // - mediaKeyMessageError (only fires under 'might not work' circumstances)
-    this.retriggerError_ = (event) => {
+    this.retriggerError_ = function (event) {
       if (event.error === 'capability' && event.event === 'mediasource') {
         // No support for MSE
-        this.player.error({
+        _this.player.error({
           code: 4,
-          message: 'The media cannot be played because it requires a feature ' +
-            'that your browser does not support.'
+          message: 'The media cannot be played because it requires a feature ' + 'that your browser does not support.'
         });
-
-      } else if (event.error === 'manifestError' && (
-        // Manifest type not supported
-        (event.event.id === 'createParser') ||
-        // Codec(s) not supported
-        (event.event.id === 'codec') ||
-        // No streams available to stream
-        (event.event.id === 'nostreams') ||
-        // Error creating Stream object
-        (event.event.id === 'nostreamscomposed') ||
-        // syntax error parsing the manifest
-        (event.event.id === 'parse') ||
-        // a stream has multiplexed audio+video
-        (event.event.id === 'multiplexedrep')
+      } else if (event.error === 'manifestError' && (event.event.id === 'createParser' || // Manifest type not supported
+      event.event.id === 'codec' || // Codec(s) not supported
+      event.event.id === 'nostreams' || // No streams available to stream
+      event.event.id === 'nostreamscomposed' || // Error creating Stream object
+      event.event.id === 'parse' || // syntax error parsing the manifest
+      event.event.id === 'multiplexedrep' // a stream has multiplexed audio+video
       )) {
         // These errors have useful error messages, so we forward it on
-        this.player.error({code: 4, message: event.event.message});
-
+        _this.player.error({ code: 4, message: event.event.message });
       } else if (event.error === 'mediasource') {
         // This error happens when dash.js fails to allocate a SourceBuffer
         // OR the underlying video element throws a `MediaError`.
@@ -127,56 +132,44 @@ class Html5DashJS {
         // If it's a `MediaError`, dash.js inspects the error object for
         // additional information to append to the error type.
         if (event.event.match('MEDIA_ERR_ABORTED')) {
-          this.player.error({code: 1, message: event.event});
+          _this.player.error({ code: 1, message: event.event });
         } else if (event.event.match('MEDIA_ERR_NETWORK')) {
-          this.player.error({code: 2, message: event.event});
+          _this.player.error({ code: 2, message: event.event });
         } else if (event.event.match('MEDIA_ERR_DECODE')) {
-          this.player.error({code: 3, message: event.event});
+          _this.player.error({ code: 3, message: event.event });
         } else if (event.event.match('MEDIA_ERR_SRC_NOT_SUPPORTED')) {
-          this.player.error({code: 4, message: event.event});
+          _this.player.error({ code: 4, message: event.event });
         } else if (event.event.match('MEDIA_ERR_ENCRYPTED')) {
-          this.player.error({code: 5, message: event.event});
+          _this.player.error({ code: 5, message: event.event });
         } else if (event.event.match('UNKNOWN')) {
           // We shouldn't ever end up here, since this would mean a
           // `MediaError` thrown by the video element that doesn't comply
           // with the W3C spec. But, since we should handle the error,
           // throwing a MEDIA_ERR_SRC_NOT_SUPPORTED is probably the
           // most reasonable thing to do.
-          this.player.error({code: 4, message: event.event});
+          _this.player.error({ code: 4, message: event.event });
         } else {
           // Buffer allocation error
-          this.player.error({code: 4, message: event.event});
+          _this.player.error({ code: 4, message: event.event });
         }
-
       } else if (event.error === 'capability' && event.event === 'encryptedmedia') {
         // Browser doesn't support EME
-        this.player.error({
+        _this.player.error({
           code: 5,
-          message: 'The media cannot be played because it requires encryption ' +
-            'features that your browser does not support.'
+          message: 'The media cannot be played because it requires encryption ' + 'features that your browser does not support.'
         });
-
       } else if (event.error === 'key_session') {
         // This block handles pretty much all errors thrown by the
         // encryption subsystem
-        this.player.error({
+        _this.player.error({
           code: 5,
           message: event.event
         });
-
       } else if (event.error === 'download') {
-        this.player.error({
+        _this.player.error({
           code: 2,
-          message: 'The media playback was aborted because too many consecutive ' +
-            'download errors occurred.'
+          message: 'The media playback was aborted because too many consecutive ' + 'download errors occurred.'
         });
-
-      } else if (event.error === 'mssError') {
-        this.player.error({
-          code: 3,
-          message: event.event
-        });
-
       } else {
         // ignore the error
         return;
@@ -184,52 +177,31 @@ class Html5DashJS {
 
       // only reset the dash player in 10ms async, so that the rest of the
       // calling function finishes
-      setTimeout(() => {
-        this.mediaPlayer_.reset();
+      setTimeout(function () {
+        _this.mediaPlayer_.reset();
       }, 10);
     };
 
-    this.mediaPlayer_.on(dashjs.MediaPlayer.events.ERROR, this.retriggerError_);
-    this.mediaPlayer_.on(dashjs.MediaPlayer.events.PLAYBACK_TIME_UPDATED, this.timeUpdated);
-
-    this.getDuration_ = (event) => {
-      const periods = event.data.Period_asArray;
-      const oldHasFiniteDuration = this.hasFiniteDuration_;
-
-      if (event.data.mediaPresentationDuration || periods[periods.length - 1].duration) {
-        this.hasFiniteDuration_ = true;
-      } else {
-        // in case we run into a weird situation where we're VOD but then
-        // switch to live
-        this.hasFiniteDuration_ = false;
-      }
-
-      if (this.hasFiniteDuration_ !== oldHasFiniteDuration) {
-        this.player.trigger('durationchange');
-      }
-    };
-
-    this.mediaPlayer_.on(dashjs.MediaPlayer.events.MANIFEST_LOADED, this.getDuration_);
-
+    this.mediaPlayer_.on(_dashjs2['default'].MediaPlayer.events.ERROR, this.retriggerError_);
+    this.mediaPlayer_.on(_dashjs2['default'].MediaPlayer.events.PLAYBACK_TIME_UPDATED, this.timeUpdated);
     // Apply all dash options that are set
     if (options.dash) {
-      Object.keys(options.dash).forEach((key) => {
-        const dashOptionsKey = 'set' + key.charAt(0).toUpperCase() + key.slice(1);
-        let value = options.dash[key];
+      Object.keys(options.dash).forEach(function (key) {
+        var _mediaPlayer_;
 
-        if (this.mediaPlayer_.hasOwnProperty(dashOptionsKey)) {
+        var dashOptionsKey = 'set' + key.charAt(0).toUpperCase() + key.slice(1);
+        var value = options.dash[key];
+
+        if (_this.mediaPlayer_.hasOwnProperty(dashOptionsKey)) {
           // Providing a key without `set` prefix is now deprecated.
-          videojs.log.warn('Using dash options in videojs-contrib-dash without the set prefix ' +
-            `has been deprecated. Change '${key}' to '${dashOptionsKey}'`);
+          _video2['default'].log.warn('Using dash options in videojs-contrib-dash without the set prefix ' + ('has been deprecated. Change \'' + key + '\' to \'' + dashOptionsKey + '\''));
 
           // Set key so it will still work
           key = dashOptionsKey;
         }
 
-        if (!this.mediaPlayer_.hasOwnProperty(key)) {
-          videojs.log.warn(
-            `Warning: dash configuration option unrecognized: ${key}`
-          );
+        if (!_this.mediaPlayer_.hasOwnProperty(key)) {
+          _video2['default'].log.warn('Warning: dash configuration option unrecognized: ' + key);
 
           return;
         }
@@ -239,36 +211,34 @@ class Html5DashJS {
           value = [value];
         }
 
-        this.mediaPlayer_[key](...value);
+        (_mediaPlayer_ = _this.mediaPlayer_)[key].apply(_mediaPlayer_, value);
       });
     }
 
     this.mediaPlayer_.attachView(this.el_);
-    
+
     // Set autoplay to be the videojs autoplay setting, 
     // making sure that any truthy values are considered true
     this.mediaPlayer_.setAutoPlay(!!this.player.autoplay());
 
     // Setup audio tracks
-    setupAudioTracks.call(null, this.player, tech);
+    _setupAudioTracks2['default'].call(null, this.player, tech);
 
     // Setup text tracks
-    setupTextTracks.call(null, this.player, tech, options);
+    _setupTextTracks2['default'].call(null, this.player, tech, options);
 
     // Attach the source with any protection data
     this.mediaPlayer_.setProtectionData(this.keySystemOptions_);
     this.mediaPlayer_.attachSource(manifestSource);
-		this.mediaPlayer_.attachTTMLRenderingDiv(
-			this.player.textTrackDisplay.el_
-    );
+    this.mediaPlayer_.attachTTMLRenderingDiv(this.player.textTrackDisplay.el_);
 
     this.timeOffset = 0.0;
     this.pastSeekEnd = 0.0;
-    let that = this;
-    this.mediaPlayer_.on("manifestLoaded", e => {
+    var that = this;
+    this.mediaPlayer_.on("manifestLoaded", function (e) {
       // If we get a manifest update for each segment then we
       // update timeOffset (seekable range) and reset pastSeekEnd
-      let tempPastSeekEnd = that.pastSeekEnd;
+      var tempPastSeekEnd = that.pastSeekEnd;
       that.pastSeekEnd = 0.0;
       if (e.data.type === "dynamic") {
         // this.timeOffset += e.data.minBufferTime;
@@ -276,7 +246,7 @@ class Html5DashJS {
         if (that.pastSeekEndInterval) {
           clearInterval(that.pastSeekEndInterval);
         }
-        that.pastSeekEndInterval = setInterval(() => {
+        that.pastSeekEndInterval = setInterval(function () {
           that.pastSeekEnd += 0.03;
 
           // For number based live manifests we might not get frequent manifest updates
@@ -298,16 +268,18 @@ class Html5DashJS {
    *
    * Also rename 'licenseUrl' property in the options to an 'serverURL' property
    */
-  static buildDashJSProtData(keySystemOptions) {
-    const output = {};
+
+
+  Html5DashJS.buildDashJSProtData = function buildDashJSProtData(keySystemOptions) {
+    var output = {};
 
     if (!keySystemOptions || !Array.isArray(keySystemOptions)) {
       return null;
     }
 
-    for (let i = 0; i < keySystemOptions.length; i++) {
-      const keySystem = keySystemOptions[i];
-      const options = videojs.mergeOptions({}, keySystem.options);
+    for (var i = 0; i < keySystemOptions.length; i++) {
+      var keySystem = keySystemOptions[i];
+      var options = _video2['default'].mergeOptions({}, keySystem.options);
 
       if (options.licenseUrl) {
         options.serverURL = options.licenseUrl;
@@ -318,48 +290,52 @@ class Html5DashJS {
     }
 
     return output;
-  }
+  };
 
-  dispose() {
+  Html5DashJS.prototype.dispose = function dispose() {
     if (this.mediaPlayer_) {
-      this.mediaPlayer_.off(dashjs.MediaPlayer.events.PLAYBACK_TIME_UPDATED, this.timeUpdated);
-      this.mediaPlayer_.off(dashjs.MediaPlayer.events.ERROR, this.retriggerError_);
-      this.mediaPlayer_.off(dashjs.MediaPlayer.events.MANIFEST_LOADED, this.getDuration_);
+      this.mediaPlayer_.off(_dashjs2['default'].MediaPlayer.events.PLAYBACK_TIME_UPDATED, this.timeUpdated);
+      this.mediaPlayer_.off(_dashjs2['default'].MediaPlayer.events.ERROR, this.retriggerError_);
       this.mediaPlayer_.reset();
     }
     if (this.pastSeekEndInterval) {
       clearInterval(this.pastSeekEndInterval);
     }
 
-
     if (this.player.dash) {
       delete this.player.dash;
     }
-  }
+  };
 
-  duration() {
-    if (this.mediaPlayer_.isDynamic() && !this.hasFiniteDuration_) {
-      return Infinity;
+  Html5DashJS.prototype.duration = function duration() {
+    var duration = this.mediaPlayer_.duration();
+    if (this.mediaPlayer_.isDynamic()) {
+      return -duration;
     }
-    return this.mediaPlayer_.duration();
+    return duration;
+  };
 
-  }
-
-  currentTime() {
+  Html5DashJS.prototype.currentTime = function currentTime() {
     var time = this.mediaPlayer_.time();
-    let result = time + this.timeOffset + this.pastSeekEnd;
+    var result = time + this.timeOffset + this.pastSeekEnd;
     return result;
   };
 
-  setCurrentTime(seekTime) {
-    let time = seekTime - this.timeOffset - this.pastSeekEnd;
+  Html5DashJS.prototype.setCurrentTime = function setCurrentTime(seekTime) {
+    var time = seekTime - this.timeOffset - this.pastSeekEnd;
     this.mediaPlayer_.seek(time);
-  }
+  };
 
-  seekable() {
-    let duration = this.mediaPlayer_.duration();
-    return { length: 1, start: () => this.timeOffset, end: () => this.timeOffset + duration };
-  }
+  Html5DashJS.prototype.seekable = function seekable() {
+    var _this2 = this;
+
+    var duration = this.mediaPlayer_.duration();
+    return { length: 1, start: function start() {
+        return _this2.timeOffset;
+      }, end: function end() {
+        return _this2.timeOffset + duration;
+      } };
+  };
 
   /**
    * Get a list of hooks for a specific lifecycle
@@ -369,7 +345,9 @@ class Html5DashJS {
    * @return {Array} an array of hooks or epty if none
    * @method hooks
    */
-  static hooks(type, hook) {
+
+
+  Html5DashJS.hooks = function hooks(type, hook) {
     Html5DashJS.hooks_[type] = Html5DashJS.hooks_[type] || [];
 
     if (hook) {
@@ -377,7 +355,7 @@ class Html5DashJS {
     }
 
     return Html5DashJS.hooks_[type];
-  }
+  };
 
   /**
    * Add a function hook to a specific dash lifecycle
@@ -386,9 +364,11 @@ class Html5DashJS {
    * @param {Function|Function[]} hook the function or array of functions to attach
    * @method hook
    */
-  static hook(type, hook) {
-    Html5DashJS.hooks(type, hook);
-  }
+
+
+  Html5DashJS.hook = function hook(type, _hook) {
+    Html5DashJS.hooks(type, _hook);
+  };
 
   /**
    * Remove a hook from a specific dash lifecycle.
@@ -398,8 +378,10 @@ class Html5DashJS {
    * @return {boolean} True if the function was removed, false if not found
    * @method removeHook
    */
-  static removeHook(type, hook) {
-    const index = Html5DashJS.hooks(type).indexOf(hook);
+
+
+  Html5DashJS.removeHook = function removeHook(type, hook) {
+    var index = Html5DashJS.hooks(type).indexOf(hook);
 
     if (index === -1) {
       return false;
@@ -409,69 +391,67 @@ class Html5DashJS {
     Html5DashJS.hooks_[type].splice(index, 1);
 
     return true;
-  }
-}
+  };
+
+  return Html5DashJS;
+}();
 
 Html5DashJS.hooks_ = {};
 
-const canHandleKeySystems = function(source) {
+var canHandleKeySystems = function canHandleKeySystems(source) {
   // copy the source
   source = JSON.parse(JSON.stringify(source));
 
   if (Html5DashJS.updateSourceData) {
-    videojs.log.warn('updateSourceData has been deprecated.' +
-      ' Please switch to using hook("updatesource", callback).');
+    _video2['default'].log.warn('updateSourceData has been deprecated.' + ' Please switch to using hook("updatesource", callback).');
     source = Html5DashJS.updateSourceData(source);
   }
 
   // call updatesource hooks
-  Html5DashJS.hooks('updatesource').forEach((hook) => {
+  Html5DashJS.hooks('updatesource').forEach(function (hook) {
     source = hook(source);
   });
 
-  const videoEl = document.createElement('video');
-
-  if (source.keySystemOptions &&
-    !(window.navigator.requestMediaKeySystemAccess ||
-      // IE11 Win 8.1
-      videoEl.msSetMediaKeys)) {
+  var videoEl = document.createElement('video');
+  if (source.keySystemOptions && !(navigator.requestMediaKeySystemAccess ||
+  // IE11 Win 8.1
+  videoEl.msSetMediaKeys)) {
     return false;
   }
 
   return true;
 };
 
-videojs.DashSourceHandler = function() {
+_video2['default'].DashSourceHandler = function () {
   return {
-    canHandleSource(source) {
-      const dashExtRE = /\.mpd/i;
+    canHandleSource: function canHandleSource(source) {
+      var dashExtRE = /\.mpd/i;
 
       if (!canHandleKeySystems(source)) {
         return '';
       }
 
-      if (videojs.DashSourceHandler.canPlayType(source.type)) {
+      if (_video2['default'].DashSourceHandler.canPlayType(source.type)) {
         return 'probably';
       } else if (dashExtRE.test(source.src)) {
         return 'maybe';
+      } else {
+        return '';
       }
-      return '';
-
     },
 
-    handleSource(source, tech, options) {
+    handleSource: function handleSource(source, tech, options) {
       return new Html5DashJS(source, tech, options);
     },
 
-    canPlayType(type) {
-      return videojs.DashSourceHandler.canPlayType(type);
+    canPlayType: function canPlayType(type) {
+      return _video2['default'].DashSourceHandler.canPlayType(type);
     }
   };
 };
 
-videojs.DashSourceHandler.canPlayType = function(type) {
-  const dashTypeRE = /^application\/dash\+xml/i;
-
+_video2['default'].DashSourceHandler.canPlayType = function (type) {
+  var dashTypeRE = /^application\/dash\+xml/i;
   if (dashTypeRE.test(type)) {
     return 'probably';
   }
@@ -480,9 +460,9 @@ videojs.DashSourceHandler.canPlayType = function(type) {
 };
 
 // Only add the SourceHandler if the browser supports MediaSourceExtensions
-if (window.MediaSource) {
-  videojs.getTech('Html5').registerSourceHandler(videojs.DashSourceHandler(), 0);
+if (!!_window2['default'].MediaSource) {
+  _video2['default'].getTech('Html5').registerSourceHandler(_video2['default'].DashSourceHandler(), 0);
 }
 
-videojs.Html5DashJS = Html5DashJS;
-export default Html5DashJS;
+_video2['default'].Html5DashJS = Html5DashJS;
+exports['default'] = Html5DashJS;