+++ src/js/setup-audio-tracks.js
@@ -20,8 +20,8 @@ function handlePlaybackMetadataLoaded(player, tech) {
     return `dash-audio-${index}`;
   }
 
-  function findDashAudioTrack(dashAudioTracks, videojsAudioTrack) {
-    return dashAudioTracks.find(({index}) =>
+  function findDashAudioTrack(subDashAudioTracks, videojsAudioTrack) {
+    return subDashAudioTracks.find(({index}) =>
       generateIdFromTrackIndex(index) === videojsAudioTrack.id
     );
   }
@@ -35,10 +35,33 @@ function handlePlaybackMetadataLoaded(player, tech) {
   const currentAudioTrack = mediaPlayer.getCurrentTrackFor('audio');
 
   dashAudioTracks.forEach((dashTrack) => {
-    let label = dashTrack.lang;
+    let localizedLabel;
 
-    if (dashTrack.roles && dashTrack.roles.length) {
-      label += ` (${dashTrack.roles.join(', ')})`;
+    if (Array.isArray(dashTrack.labels)) {
+      for (let i = 0; i < dashTrack.labels.length; i++) {
+        if (
+          dashTrack.labels[i].lang &&
+          player.language().indexOf(dashTrack.labels[i].lang.toLowerCase()) !== -1
+        ) {
+          localizedLabel = dashTrack.labels[i];
+
+          break;
+        }
+      }
+    }
+
+    let label;
+
+    if (localizedLabel) {
+      label = localizedLabel.text;
+    } else if (Array.isArray(dashTrack.labels) && dashTrack.labels.length === 1) {
+      label = dashTrack.labels[0].text;
+    } else {
+      label = dashTrack.lang;
+
+      if (dashTrack.roles && dashTrack.roles.length) {
+        label += ' (' + dashTrack.roles.join(', ') + ')';
+      }
     }
 
     // Add the track to the player's audio track list.
@@ -48,19 +71,18 @@ function handlePlaybackMetadataLoaded(player, tech) {
         id: generateIdFromTrackIndex(dashTrack.index),
         kind: dashTrack.kind || 'main',
         label,
-        language: dashTrack.lang,
+        language: dashTrack.lang
       })
     );
   });
 
-  let audioTracksChangeHandler = () => {
+  const audioTracksChangeHandler = () => {
     for (let i = 0; i < videojsAudioTracks.length; i++) {
       const track = videojsAudioTracks[i];
 
       if (track.enabled) {
         // Find the audio track we just selected by the id
         const dashAudioTrack = findDashAudioTrack(dashAudioTracks, track);
-
 
         // Set is as the current track
         mediaPlayer.setCurrentTrack(dashAudioTrack);
@@ -84,7 +106,7 @@ function handlePlaybackMetadataLoaded(player, tech) {
 export default function setupAudioTracks(player, tech) {
   // When `dashjs` finishes loading metadata, create audio tracks for `video.js`.
   player.dash.mediaPlayer.on(
-      dashjs.MediaPlayer.events.PLAYBACK_METADATA_LOADED,
-      handlePlaybackMetadataLoaded.bind(null, player, tech)
+    dashjs.MediaPlayer.events.PLAYBACK_METADATA_LOADED,
+    handlePlaybackMetadataLoaded.bind(null, player, tech)
   );
 }